src/requirements.txt
asgiref==3.4.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
certifi==2021.10.8; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.15.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
charset-normalizer==2.0.7; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3"
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
cryptography==35.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
django-cors-headers==3.10.0; python_version >= "3.6"
django-filter==21.1; python_version >= "3.6"
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==3.2.8; python_version >= "3.6"
djangorestframework-simplejwt==4.8.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
drf-yasg==1.20.0; python_version >= "3.6"
idna==3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.5"
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jinja2==3.0.2; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
oauthlib==3.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
packaging==21.0; python_version >= "3.6"
pillow==8.4.0; python_version >= "3.6"
psycopg2==2.9.1; python_version >= "3.6"
pycparser==2.20; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyjwt==2.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2021.3; python_version >= "3.6"
requests-oauthlib==1.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.26.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.10" and python_version >= "3.6"
ruamel.yaml==0.17.16; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
uritemplate==4.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
urllib3==1.26.7; python_full_version >= "3.6.1" and python_version < "4" and python_full_version < "4.0.0"
